Get rid of the 'type inference' module
Almost all of our type inference is done inline as part of `compile`,
most AST elements have a `typeref` field in them. So for most AST nodes,
"type inference" simply converts the typeref to a schema type.

There are a handful of nodes that do non-trivial typechecking work as
part of type inference, but there really isn't any benefit to us to
having them there, and there is some cost in needing to have the
infrastructure.

Instead we require every `Expr` to have a `typeref`, and we move the
typechecking to the compile sites for the few cases where it is
relevant.

(The thing that prompted this was Victor discovering the nearly-dead
`_infer_binop_args` function that seemed like something he cared about
but was in fact only used in a totally trivial way.)
