+++ edb/edgeql/compiler/expr.py
@@ -27,6 +27,7 @@ from typing import *
 from edb import errors
 
 from edb.common import context as ctx_utils
+from edb.common import parsing
 
 from edb.ir import ast as irast
 from edb.ir import typeutils as irtyputils
@@ -38,6 +39,8 @@ from edb.schema import globals as s_globals
 from edb.schema import indexes as s_indexes
 from edb.schema import name as sn
 from edb.schema import objtypes as s_objtypes
+from edb.schema import pseudo as s_pseudo
+from edb.schema import scalars as s_scalars
 from edb.schema import types as s_types
 from edb.schema import utils as s_utils
 
@@ -48,7 +51,6 @@ from . import astutils
 from . import casts
 from . import context
 from . import dispatch
-from . import inference
 from . import pathctx
 from . import setgen
 from . import stmt
@@ -302,7 +304,7 @@ def compile_Array(
     ]
     # check that none of the elements are themselves arrays
     for el, expr_el in zip(elements, expr.elements):
-        if isinstance(inference.infer_type(el, ctx.env), s_abc.Array):
+        if isinstance(setgen.get_set_type(el, ctx=ctx), s_abc.Array):
             raise errors.QueryError(
                 f'nested arrays are not supported',
                 context=expr_el.context)
@@ -689,6 +691,34 @@ def compile_TypeCast(
     return stmt.maybe_add_view(res, ctx=ctx)
 
 
+def _infer_type_introspection(
+    typeref: irast.TypeRef,
+    env: context.Environment,
+    srcctx: Optional[parsing.ParserContext]=None,
+) -> s_types.Type:
+    if irtyputils.is_scalar(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::ScalarType'))
+    elif irtyputils.is_object(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::ObjectType'))
+    elif irtyputils.is_array(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::Array'))
+    elif irtyputils.is_tuple(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::Tuple'))
+    elif irtyputils.is_range(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::Range'))
+    elif irtyputils.is_multirange(typeref):
+        return cast(s_objtypes.ObjectType,
+                    env.schema.get('schema::MultiRange'))
+    else:
+        raise errors.QueryError(
+            'unexpected type in INTROSPECT', context=srcctx)
+
+
 @dispatch.compile.register(qlast.Introspect)
 def compile_Introspect(
         expr: qlast.Introspect, *, ctx: context.ContextLevel) -> irast.Set:
@@ -718,8 +748,137 @@ def compile_Introspect(
             f'cannot introspect generic types',
             context=expr.type.context)
 
-    ir = setgen.ensure_set(irast.TypeIntrospection(typeref=typeref), ctx=ctx)
+    result_typeref = typegen.type_to_typeref(
+        _infer_type_introspection(typeref, ctx.env, expr.context), env=ctx.env
+    )
+    ir = setgen.ensure_set(
+        irast.TypeIntrospection(output_typeref=typeref, typeref=result_typeref),
+        ctx=ctx,
+    )
     return stmt.maybe_add_view(ir, ctx=ctx)
+
+
+def _infer_index_type(
+    expr: irast.Set | irast.Expr,
+    index: irast.Set,
+    *, ctx: context.ContextLevel,
+) -> s_types.Type:
+    env = ctx.env
+    node_type = setgen.get_expr_type(expr, ctx=ctx)
+    index_type = setgen.get_set_type(index, ctx=ctx)
+
+    str_t = env.schema.get('std::str', type=s_scalars.ScalarType)
+    bytes_t = env.schema.get('std::bytes', type=s_scalars.ScalarType)
+    int_t = env.schema.get('std::int64', type=s_scalars.ScalarType)
+    json_t = env.schema.get('std::json', type=s_scalars.ScalarType)
+
+    result: s_types.Type
+
+    if node_type.issubclass(env.schema, str_t):
+
+        if not index_type.implicitly_castable_to(int_t, env.schema):
+            raise errors.QueryError(
+                f'cannot index string by '
+                f'{index_type.get_displayname(env.schema)}, '
+                f'{int_t.get_displayname(env.schema)} was expected',
+                context=index.context)
+
+        result = str_t
+
+    elif node_type.issubclass(env.schema, bytes_t):
+
+        if not index_type.implicitly_castable_to(int_t, env.schema):
+            raise errors.QueryError(
+                f'cannot index bytes by '
+                f'{index_type.get_displayname(env.schema)}, '
+                f'{int_t.get_displayname(env.schema)} was expected',
+                context=index.context)
+
+        result = bytes_t
+
+    elif node_type.issubclass(env.schema, json_t):
+
+        if not (index_type.implicitly_castable_to(int_t, env.schema) or
+                index_type.implicitly_castable_to(str_t, env.schema)):
+
+            raise errors.QueryError(
+                f'cannot index json by '
+                f'{index_type.get_displayname(env.schema)}, '
+                f'{int_t.get_displayname(env.schema)} or '
+                f'{str_t.get_displayname(env.schema)} was expected',
+                context=index.context)
+
+        result = json_t
+
+    elif isinstance(node_type, s_types.Array):
+
+        if not index_type.implicitly_castable_to(int_t, env.schema):
+            raise errors.QueryError(
+                f'cannot index array by '
+                f'{index_type.get_displayname(env.schema)}, '
+                f'{int_t.get_displayname(env.schema)} was expected',
+                context=index.context)
+
+        result = node_type.get_subtypes(env.schema)[0]
+
+    elif (node_type.is_any(env.schema) or
+            (node_type.is_scalar() and
+                str(node_type.get_name(env.schema)) == 'std::anyscalar') and
+            (index_type.implicitly_castable_to(int_t, env.schema) or
+                index_type.implicitly_castable_to(str_t, env.schema))):
+        result = s_pseudo.PseudoType.get(env.schema, 'anytype')
+
+    else:
+        raise errors.QueryError(
+            f'index indirection cannot be applied to '
+            f'{node_type.get_verbosename(env.schema)}',
+            context=expr.context)
+
+    return result
+
+
+def _infer_slice_type(
+    expr: irast.Set,
+    start: Optional[irast.Set],
+    stop: Optional[irast.Set],
+    *, ctx: context.ContextLevel,
+) -> s_types.Type:
+    env = ctx.env
+    node_type = setgen.get_set_type(expr, ctx=ctx)
+
+    str_t = env.schema.get('std::str', type=s_scalars.ScalarType)
+    int_t = env.schema.get('std::int64', type=s_scalars.ScalarType)
+    json_t = env.schema.get('std::json', type=s_scalars.ScalarType)
+    bytes_t = env.schema.get('std::bytes', type=s_scalars.ScalarType)
+
+    if node_type.issubclass(env.schema, str_t):
+        base_name = 'string'
+    elif node_type.issubclass(env.schema, json_t):
+        base_name = 'JSON array'
+    elif node_type.issubclass(env.schema, bytes_t):
+        base_name = 'bytes'
+    elif isinstance(node_type, s_abc.Array):
+        base_name = 'array'
+    elif node_type.is_any(env.schema):
+        base_name = 'anytype'
+    else:
+        # the base type is not valid
+        raise errors.QueryError(
+            f'{node_type.get_verbosename(env.schema)} cannot be sliced',
+            context=expr.context)
+
+    for index in [start, stop]:
+        if index is not None:
+            index_type = setgen.get_set_type(index, ctx=ctx)
+
+            if not index_type.implicitly_castable_to(int_t, env.schema):
+                raise errors.QueryError(
+                    f'cannot slice {base_name} by '
+                    f'{index_type.get_displayname(env.schema)}, '
+                    f'{int_t.get_displayname(env.schema)} was expected',
+                    context=index.context)
+
+    return node_type
 
 
 @dispatch.compile.register(qlast.Indirection)
@@ -731,10 +890,13 @@ def compile_Indirection(
         if isinstance(indirection_el, qlast.Index):
             idx = dispatch.compile(indirection_el.index, ctx=ctx)
             idx.context = indirection_el.index.context
-            node = irast.IndexIndirection(
-                expr=node, index=idx, context=expr.context
+            typeref = typegen.type_to_typeref(
+                _infer_index_type(node, idx, ctx=ctx), env=ctx.env
             )
 
+            node = irast.IndexIndirection(
+                expr=node, index=idx, typeref=typeref, context=expr.context
+            )
         elif isinstance(indirection_el, qlast.Slice):
             start: Optional[irast.Base]
             stop: Optional[irast.Base]
@@ -750,8 +912,11 @@ def compile_Indirection(
                 stop = None
 
             node_set = setgen.ensure_set(node, ctx=ctx)
+            typeref = typegen.type_to_typeref(
+                _infer_slice_type(node_set, start, stop, ctx=ctx), env=ctx.env
+            )
             node = irast.SliceIndirection(
-                expr=node_set, start=start, stop=stop
+                expr=node_set, start=start, stop=stop, typeref=typeref,
             )
         else:
             raise ValueError(
@@ -787,8 +952,14 @@ def compile_type_check_op(
         )
         result = ltype.issubclass(ctx.env.schema, test_type)
 
+    output_typeref = typegen.type_to_typeref(
+        ctx.env.schema.get('std::bool', type=s_types.Type),
+        env=ctx.env,
+    )
+
     return irast.TypeCheckOp(
-        left=left, right=typeref, op=expr.op, result=result)
+        left=left, right=typeref, op=expr.op, result=result,
+        typeref=output_typeref)
 
 
 def flatten_set(expr: qlast.Set) -> List[qlast.Expr]:
