+++ edb/edgeql/compiler/setgen.py
@@ -130,6 +130,12 @@ def get_set_type(
     return ctx.env.set_types[ir_set]
 
 
+def get_expr_type(
+        ir: irast.Set | irast.Expr, *,
+        ctx: context.ContextLevel) -> s_types.Type:
+    return typegen.type_from_typeref(ir.typeref, env=ctx.env)
+
+
 class KeepCurrentT(enum.Enum):
     KeepCurrent = 0
 
@@ -206,9 +212,9 @@ def new_tuple_set(
         named: bool,
         ctx: context.ContextLevel) -> irast.Set:
 
-    dummy_typeref = cast(irast.TypeRef, None)
-    tup = irast.Tuple(elements=elements, named=named, typeref=dummy_typeref)
-    stype = inference.infer_type(tup, env=ctx.env)
+    element_types = {el.name: get_set_type(el.val, ctx=ctx) for el in elements}
+    ctx.env.schema, stype = s_types.Tuple.create(
+        ctx.env.schema, element_types=element_types, named=named)
     result_path_id = pathctx.get_expression_path_id(stype, ctx=ctx)
 
     final_elems = []
@@ -223,8 +229,8 @@ def new_tuple_set(
         ))
 
     typeref = typegen.type_to_typeref(stype, env=ctx.env)
-    final_tup = irast.Tuple(elements=final_elems, named=named, typeref=typeref)
-    return ensure_set(final_tup, path_id=result_path_id,
+    tup = irast.Tuple(elements=final_elems, named=named, typeref=typeref)
+    return ensure_set(tup, path_id=result_path_id,
                       type_override=stype, ctx=ctx)
 
 
@@ -234,22 +240,25 @@ def new_array_set(
         ctx: context.ContextLevel,
         srcctx: Optional[parsing.ParserContext]=None) -> irast.Set:
 
-    dummy_typeref = cast(irast.TypeRef, None)
-    arr = irast.Array(elements=elements, typeref=dummy_typeref)
     if elements:
-        stype = inference.infer_type(arr, env=ctx.env)
-    elif stype is not None and stype.is_array():
+        element_type = typegen.infer_common_type(elements, ctx.env)
+        if element_type is None:
+            raise errors.QueryError('could not determine array type',
+                                    context=srcctx)
+    elif stype is not None:
         # When constructing an empty array, we should skip explicit cast any
         # time that we would skip it for an empty set because we can infer it
         # from the context.
-        pass
+        assert stype.is_array()
     else:
-        anytype = s_pseudo.PseudoType.get(ctx.env.schema, 'anytype')
-        ctx.env.schema, stype = s_types.Array.from_subtypes(
-            ctx.env.schema, [anytype])
+        element_type = s_pseudo.PseudoType.get(ctx.env.schema, 'anytype')
         if srcctx is not None:
-            ctx.env.type_origins[anytype] = srcctx
+            ctx.env.type_origins[element_type] = srcctx
 
+    if stype is None:
+        assert element_type
+        ctx.env.schema, stype = s_types.Array.from_subtypes(
+            ctx.env.schema, [element_type])
     typeref = typegen.type_to_typeref(stype, env=ctx.env)
     arr = irast.Array(elements=elements, typeref=typeref)
     return ensure_set(arr, type_override=stype, ctx=ctx)
@@ -479,7 +488,7 @@ def compile_path(expr: qlast.Path, *, ctx: context.ContextLevel) -> irast.Set:
                     is_computable = True
 
         elif isinstance(step, qlast.TypeIntersection):
-            arg_type = inference.infer_type(path_tip, ctx.env)
+            arg_type = get_set_type(path_tip, ctx=ctx)
             if not isinstance(arg_type, s_objtypes.ObjectType):
                 raise errors.QueryError(
                     f'cannot apply type intersection operator to '
@@ -1266,7 +1275,7 @@ def expression_set(
     if type_override is not None:
         stype = type_override
     else:
-        stype = inference.infer_type(expr, ctx.env)
+        stype = get_expr_type(expr, ctx=ctx)
 
     if path_id is None:
         path_id = getattr(expr, 'path_id', None)
@@ -1346,7 +1355,7 @@ def ensure_set(
     if (isinstance(ir_set, irast.EmptySet)
             and (stype is None or stype.is_any(ctx.env.schema))
             and typehint is not None):
-        inference.amend_empty_set_type(ir_set, typehint, env=ctx.env)
+        typegen.amend_empty_set_type(ir_set, typehint, env=ctx.env)
         stype = get_set_type(ir_set, ctx=ctx)
 
     if (
