+++ edb/ir/ast.py
@@ -65,6 +65,7 @@ Set (
 
 from __future__ import annotations
 
+import abc
 import dataclasses
 import typing
 import uuid
@@ -485,6 +486,12 @@ class TupleIndirectionPointer(Pointer):
 class Expr(Base):
     __abstract_node__ = True
 
+    if typing.TYPE_CHECKING:
+        @property
+        @abc.abstractmethod
+        def typeref(self) -> TypeRef:
+            raise NotImplementedError
+
     # Sets to materialize at this point, keyed by the type/ptr id.
     materialized_sets: typing.Optional[
         typing.Dict[uuid.UUID, MaterializedSet]] = None
@@ -729,6 +736,9 @@ class Statement(Command):
 
 class TypeIntrospection(ImmutableExpr):
 
+    # The type value to return
+    output_typeref: TypeRef
+    # The type value *of the output*
     typeref: TypeRef
 
 
@@ -837,6 +847,7 @@ class TypeCheckOp(ImmutableExpr):
     right: TypeRef
     op: str
     result: typing.Optional[bool] = None
+    typeref: TypeRef
 
 
 class SortExpr(Base):
@@ -974,6 +985,7 @@ class IndexIndirection(ImmutableExpr):
 
     expr: Base
     index: Base
+    typeref: TypeRef
 
 
 class SliceIndirection(ImmutableExpr):
@@ -981,6 +993,7 @@ class SliceIndirection(ImmutableExpr):
     expr: Set
     start: typing.Optional[Base]
     stop: typing.Optional[Base]
+    typeref: TypeRef
 
 
 class TypeCast(ImmutableExpr):
@@ -994,6 +1007,10 @@ class TypeCast(ImmutableExpr):
     sql_function: typing.Optional[str] = None
     sql_cast: bool
     sql_expr: bool
+
+    @property
+    def typeref(self) -> TypeRef:
+        return self.to_type
 
 
 class MaterializedSet(Base):
@@ -1039,6 +1056,10 @@ class Stmt(Expr):
     parent_stmt: typing.Optional[Stmt] = None
     iterator_stmt: typing.Optional[Set] = None
     bindings: typing.Optional[typing.List[Set]] = None
+
+    @property
+    def typeref(self) -> TypeRef:
+        return self.result.typeref
 
 
 class FilteredStmt(Stmt):
@@ -1168,10 +1189,15 @@ class OnConflictClause(Base):
 
 class InsertStmt(MutatingStmt):
     on_conflict: typing.Optional[OnConflictClause] = None
+    final_typeref: typing.Optional[TypeRef] = None
 
     @property
     def material_type(self) -> TypeRef:
         return self.subject.typeref.real_material_type
+
+    @property
+    def typeref(self) -> TypeRef:
+        return self.final_typeref or self.result.typeref
 
 
 # N.B: The PointerRef corresponds to the *definition* point of the rewrite.
@@ -1243,18 +1269,33 @@ class ConfigSet(ConfigCommand):
     required: bool
     backend_expr: typing.Optional[Set] = None
 
+    @property
+    def typeref(self) -> TypeRef:
+        return self.expr.typeref
+
 
 class ConfigReset(ConfigCommand):
 
     selector: typing.Optional[Set] = None
+
+    @property
+    def typeref(self) -> TypeRef:
+        return TypeRef(
+            id=so.get_known_type_id('anytype'),
+            name_hint=sn.UnqualName('anytype'),
+        )
 
 
 class ConfigInsert(ConfigCommand):
 
     expr: Set
 
+    @property
+    def typeref(self) -> TypeRef:
+        return self.expr.typeref
 
-class FTSDocument(Expr):
+
+class FTSDocument(ImmutableExpr):
     """
     Text and information on how to search through it.
 
@@ -1268,3 +1309,5 @@ class FTSDocument(Expr):
     language_domain: typing.Set[str]
 
     weight: typing.Optional[str]
+
+    typeref: TypeRef
